Some errors have been fixed
