+++ java/ArrayUtilsTest.java
@@ -1,6 +1,7 @@
 package algorithms.utils;
 
 import static org.junit.jupiter.api.Assertions.*;
+import java.util.*;
 import org.junit.jupiter.api.Test;
 
 class ArrayUtilsTest {
