--- java/Climbing Stairs.java
@@ -1,17 +1,0 @@
-public class ClimbingStairs {
-    public int climbStairs(int n) {
-        if (n <= 2) return n;
-        int first = 1, second = 2;
-        for (int i = 3; i <= n; i++) {
-            int temp = first + second;
-            first = second;
-            second = temp;
-        }
-        return second;
-    }
-
-    public static void main(String[] args) {
-        ClimbingStairs obj = new ClimbingStairs();
-        System.out.println(obj.climbStairs(5)); // 8
-    }
-}
